+++ src/index.js
@@ -22,4 +22,4 @@ export { default as PSACheckbox } from "./Checkbox/index";
 export { default as PSAProfile } from "./Profile/index";
 export { default as PSARadiobutton } from "./Radiobutton/index";
 export { default as PSALoadingspinner } from "./Loadingspinner/index";
-export { default as PSAStepper } from "./Progressstepper/index";
+export { default as PSAStepper } from "./Stepper/index";
